+++ lib/src/client.dart
@@ -7,9 +7,7 @@ class Client {
   static final Uri defaultEndPoint = Uri.parse('https://smsapi.free-mobile.fr');
 
   /// Creates a new client.
-  Client([this.username, this.password, endPoint]) {
-    if (endPoint != null) this.endPoint = endPoint is Uri ? endPoint : Uri.parse(endPoint.toString());
-  }
+  Client(this.username, this.password, {Uri endPoint}): endPoint = endPoint ?? defaultEndPoint;
 
   /// The stream of "request" events.
   Stream<http.Request> get onRequest => _onRequest.stream;
@@ -18,13 +16,13 @@ class Client {
   Stream<http.Response> get onResponse => _onResponse.stream;
 
   /// The URL of the API end point.
-  Uri endPoint = defaultEndPoint;
+  final Uri endPoint;
 
   /// The identification key associated to the account.
-  String password;
+  final String password;
 
   /// The user name associated to the account.
-  String username;
+  final String username;
 
   /// The handler of "request" events.
   final StreamController<http.Request> _onRequest = new StreamController<http.Request>.broadcast();
@@ -33,9 +31,9 @@ class Client {
   final StreamController<http.Response> _onResponse = new StreamController<http.Response>.broadcast();
 
   /// Sends a SMS message to the underlying account.
+  /// Throws an [ArgumentError] if the account credentials are invalid or the specified message is empty.
   Future sendMessage(String text) async {
-    if (username == null || username.isEmpty) throw new ArgumentError('The account username is empty.');
-    if (password == null || password.isEmpty) throw new ArgumentError('The account password is empty.');
+    if (username.isEmpty || password.isEmpty) throw new ArgumentError('The account credentials are invalid.');
 
     var message = text.trim();
     if (message.isEmpty) throw new ArgumentError('The specified message is empty.');