+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 This file contains highlights of what changes on each version of the [Free Mobile for Dart](https://github.com/cedx/free-mobile.dart) library.
 
 ## Version 3.0.0
+- Breaking change: changed the signature of the constructor.
 - Breaking change: raised the required [Dart](https://www.dartlang.org) version.
 - Breaking change: the class properties are now read-only.
 - Breaking change: removed the `toJson()` and `toString()` methods.